refactor installer-code
