+++ install/hie-install.cabal
@@ -9,15 +9,22 @@ build-type:          Simple
 cabal-version:       >=2.0
 
 library
-  hs-source-dirs:      .
+  hs-source-dirs:      src
   exposed-modules:     Install
   other-modules:       BuildSystem
+                     , Stack
+                     , Version
+                     , Cabal
+                     , Print
+                     , Env
   build-depends:       base >= 4.9 && < 5
                      , shake == 0.17.3
                      , directory
                      , extra
                      , text
   default-extensions: LambdaCase
+                    , TupleSections
+  default-language:    Haskell2010
 
   if flag(run-from-stack)
     cpp-options:       -DRUN_FROM_STACK
