+++ install/src/Stack.hs
@@ -1,0 +1,60 @@
+module Stack where
+
+import           Development.Shake
+import           Development.Shake.Command
+import           Control.Monad
+
+import           Version
+import           Print
+
+-- | check `stack` has the required version
+checkStack :: Action ()
+checkStack = do
+  stackVersion <- trimmedStdout <$> execStackShake ["--numeric-version"]
+  unless (checkVersion requiredStackVersion stackVersion) $ do
+    printInStars $ stackExeIsOldFailMsg stackVersion
+    error $ stackExeIsOldFailMsg stackVersion
+
+-- | Get the local binary path of stack.
+-- Equal to the command `stack path --local-bin`
+getLocalBin :: Action FilePath
+getLocalBin = do
+  Stdout stackLocalDir' <- execStackShake ["path", "--local-bin"]
+  return $ trim stackLocalDir'
+
+
+-- | Execute a stack command for a specified ghc, discarding the output
+execStackWithGhc_ :: VersionNumber -> [String] -> Action ()
+execStackWithGhc_ versionNumber args = do
+  let stackFile = "stack-" ++ versionNumber ++ ".yaml"
+  command_ [] "stack" (("--stack-yaml=" ++ stackFile) : args)
+
+-- | Execute a stack command for a specified ghc
+execStackWithGhc :: CmdResult r => VersionNumber -> [String] -> Action r
+execStackWithGhc versionNumber args = do
+  let stackFile = "stack-" ++ versionNumber ++ ".yaml"
+  command [] "stack" (("--stack-yaml=" ++ stackFile) : args)
+
+-- | Execute a stack command with the same resolver as the build script
+execStackShake :: CmdResult r => [String] -> Action r
+execStackShake args = command [] "stack" ("--stack-yaml=shake.yaml" : args)
+
+-- | Execute a stack command with the same resolver as the build script, discarding the output
+execStackShake_ :: [String] -> Action ()
+execStackShake_ args = command_ [] "stack" ("--stack-yaml=shake.yaml" : args)
+
+
+-- | Error message when the `stack` binary is an older version
+stackExeIsOldFailMsg :: String -> String
+stackExeIsOldFailMsg stackVersion =
+  "The `stack` executable is outdated.\n"
+    ++ "found version is `"
+    ++ stackVersion
+    ++ "`.\n"
+    ++ "required version is `"
+    ++ versionToString requiredStackVersion
+    ++ "`.\n"
+    ++ "Please run `stack upgrade` to upgrade your stack installation"
+
+requiredStackVersion :: RequiredVersion
+requiredStackVersion = [1, 9, 3]
