+++ install/src/Version.hs
@@ -1,0 +1,24 @@
+module Version where
+
+import           Data.Version                             ( Version
+                                                          , parseVersion
+                                                          , makeVersion
+                                                          , showVersion
+                                                          )
+import           Text.ParserCombinators.ReadP             ( readP_to_S )
+import           Control.Monad.IO.Class
+
+
+type VersionNumber = String
+type RequiredVersion = [Int]
+
+versionToString :: RequiredVersion -> String
+versionToString = showVersion . makeVersion
+
+-- | Parse a version-string into a version. Fails if the version-string is not valid
+parseVersionEx :: String -> Version
+parseVersionEx = fst . head . filter (("" ==) . snd) . readP_to_S parseVersion
+
+
+checkVersion :: RequiredVersion -> String -> Bool
+checkVersion required given = parseVersionEx given >= makeVersion required
